refactor(audit): extract constantTimeEqual to lib/crypto-compare (PLATFORM-AUDIT PR10)
Closes BACKLOG M15-4 #15. Removes 9 identical local copies of the
timing-safe string comparison function and centralises them in
lib/crypto-compare.ts. Callers updated:

- cron/budget-reset
- cron/optimiser-monitor-rollouts
- cron/process-batch
- cron/process-brief-runner
- cron/process-regenerations
- cron/process-transfer
- emergency
- ops/reset-admin-password
- ops/self-probe

No behaviour change — same timingSafeEqual implementation, same
length-mismatch constant-time padding.
