+++ app/api/cron/process-transfer/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse, type NextRequest } from "next/server";
-import { timingSafeEqual } from "node:crypto";
 
+import { constantTimeEqual } from "@/lib/crypto-compare";
 import { logger } from "@/lib/logger";
 import {
   DEFAULT_LEASE_MS,
@@ -40,17 +40,6 @@ export const dynamic = "force-dynamic";
 // 299s keeps us just under Vercel's 300s function ceiling — belt-
 // and-suspenders for a stuck heartbeat loop.
 export const maxDuration = 299;
-
-function constantTimeEqual(a: string, b: string): boolean {
-  const aBuf = Buffer.from(a, "utf8");
-  const bBuf = Buffer.from(b, "utf8");
-  if (aBuf.length !== bBuf.length) {
-    const filler = Buffer.alloc(aBuf.length);
-    timingSafeEqual(aBuf, filler);
-    return false;
-  }
-  return timingSafeEqual(aBuf, bBuf);
-}
 
 function authorised(req: NextRequest): boolean {
   const secret = process.env.CRON_SECRET;
